scanners/domain-discovery/requirements.txt
python-arango==7.3.4
python-dotenv==0.20.0
nats-py==2.1.3